"Commting the changes for Stripe payment gateway integration"

+++ src/main/java/com/example/paymentservice/paymentgateways/IPaymentGateway.java
@@ -4,6 +4,6 @@ public interface IPaymentGateway {
     // Method to create a payment link
 
 
-    String getPaymentLink(double amount, String orderId,
+    String getPaymentLink(Long amount, String orderId,
                           String customerName, String customerContact, String customerEmail);
 }
